+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### ADDED
+- `createUser` creates a new user account using the `/api/admin/user` endpoint.
 
 ## [2.0.2] - 2022-11-07
 ### FIXED