+++ CHANGELOG.md
@@ -4,8 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### ADDED
-- Added support for the WideSky `createUser` endpoint, `/api/admin/user` for creating a new user account through one of the
-  supported local/scram authentication methods.
+- Added support for the WideSky `createUser` endpoint, `/api/admin/user`. As its name implies, this function is use for 
+  creating a new user account in WideSky through one of the supported local/scram authentication method.
 
 ## [2.0.2] - 2022-11-07
 ### FIXED